upgrade storybook version, refactor props and models
