+++ src/models/models.ts
@@ -1,0 +1,1 @@
+export type ButtonType = 'button' | 'submit' | 'reset';
